# perf: setup cfn template registry asynchronously