### Problem:

The cfn registry setup can take a long time if there are a lot of yaml files. This can significantly slow down extension startup.

### Solution:

- Have the registry setup run in the background
- If the user triggers functionality that requires the template registry and it is not yet ready, a progress message will appear on the screen

### Metrics

Testing with: https://github.com/aws/aws-sam-cli

Before: 152 seconds in MDE, 1.6 seconds local
After: 1ms in MDE, <1ms local

### Additional
If the template registry is accessed before it is done loading a message will appear to the user:
![Screen Shot 2023-04-24 at 4 15 29 PM](https://user-images.githubusercontent.com/118216176/234112340-735c80ae-98ad-4e9d-90d5-c50fdfb9e6f2.png)


## License

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.