@justinmk3 2023-04-28
Tested on a CDE 

    Failed to activate template registry. CodeLenses} will not appear on SAM template files.